+++ tests/testthat/test-read_motif_file.R
@@ -1,20 +1,28 @@
 test_that("read_motif_file can read jaspar motif file", {
     motif_file <- system.file("extdata", "motif_MA1930.2.jaspar",
                               package = "MotifPeeker")
-    # motif <- read_motif_file(motif_file, motif_id = "MA1930.2",
-    #                          file_format = "jaspar")
     motif <- check_input(motif_file, "motif", read_motif_file, inverse = TRUE,
-                        motif_id = "MA1930.2", file_format = "jaspar")
+                        file_format = "jaspar")
     expect_s4_class(motif, "universalmotif")
-    expect_error(read_motif_file(motif))
 })
 test_that("read_motif_file can infer motif file-format", {
     motif_file <- system.file("extdata", "motif_MA1930.2.jaspar",
                               package = "MotifPeeker")
-    motif <- read_motif_file(motif_file, motif_id = "MA1930.2")
+    motif <- read_motif_file(motif_file)
     expect_s4_class(motif, "universalmotif")
 })
+test_that("read_motif_file can take universalmotif object", {
+    motif_file <- system.file("extdata", "motif_MA1930.2.jaspar",
+                              package = "MotifPeeker")
+    motif <- read_motif_file(motif_file, file_format = "jaspar")
+    motif2 <- read_motif_file(motif)
+    expect_identical(motif, motif2)
+})
 test_that("read_motif_file fails with invalid data", {
-    motif_file <- system.file("extdata","CTCF_ChIP_peaks.narrowPeak")
-    expect_error(read_motif_file(motif_file, motif_id = "MA1930.2"))
+    motif_file <- system.file("extdata","CTCF_ChIP_peaks.narrowPeak",
+                              package = "MotifPeeker")
+    expect_error(read_motif_file(motif_file))
+    
+    read_file <- read_peak_file(motif_file)
+    expect_error(read_motif_file(read_file))
 })